Changes because dims[] must now be size_t
git-svn-id: https://subversion.xor.aps.anl.gov/synApps/areaDetector/trunk@15564 dc6c5ff5-0b8b-c028-a01f-ffb33f00fc8b

+++ ADApp/roperSrc/roper.cpp
@@ -240,7 +240,7 @@ NDArray *roper::getData()
     SAFEARRAY *pData;
     int dim;
     LONG lbound, ubound;
-    int dims[ND_ARRAY_MAX_DIMS];
+    size_t dims[ND_ARRAY_MAX_DIMS];
     int nDims;
     VARTYPE varType;
     NDDataType_t dataType;
@@ -313,8 +313,8 @@ NDArray *roper::getData()
         SafeArrayUnaccessData(pData);
         SafeArrayDestroy(pData);
         setIntegerParam(NDArraySize, arrayInfo.totalBytes);
-        setIntegerParam(NDArraySizeX, dims[0]);
-        setIntegerParam(NDArraySizeY, dims[1]);
+        setIntegerParam(NDArraySizeX, (int)dims[0]);
+        setIntegerParam(NDArraySizeY, (int)dims[1]);
         setIntegerParam(NDDataType, dataType);
     }
     catch(CException *pEx) {
